Move code to InsightsController
